README.md
Markov Cluster Algorithm (MCL) is a fast and scalable unsupervised graph clustering algorithm, based on stochastic flow simulation model. By random walks on the graph, the flow of the graph is estimated for the clusters. It simulates flow using two simple algebraic operations on matrices i.e. expansion and inflation. The expansion operator connects different regions of graph while inflation operator strengthens and weakens the flow. This procedure follows until a steady state is reached and all the values in column are same. This algorithm works efficiently for clusters with small diameters.  It is suitable for fast and accurate grouping of protein structures in human body. 
Restricted neighborhood search Algorithm (RNSC)   is also graph clustering technique that uses stochastic local search. Cost functions are assigned to clusters of a graph to get an optimal cost clustering. It is implemented for un-weighted and undirected graphs and uses heuristic approach. A number of moves are considered in this algorithm. A global move results in an optimal change in cost, diversification move shuffles the clustering randomly and an intensification move is associated with a good cost  chosen from restricted portion of the current clustering’s neighborhood. To escape cycling, a tabu list is created to avoid same moves in the algorithm. This algorithm is efficient as it reduces the run time of clustering algorithm. The paper proposes behavior analysis of both the algorithms effectively which is implemented in Python programming language.